+++ src/KardFinancial/Notifications/Types/EarnedRewardRejectedAttributes.cs
@@ -1,0 +1,55 @@
+using global::System.Text.Json;
+using global::System.Text.Json.Serialization;
+using KardFinancial.Core;
+
+namespace KardFinancial;
+
+[Serializable]
+public record EarnedRewardRejectedAttributes : IJsonOnDeserialized
+{
+    [JsonExtensionData]
+    private readonly IDictionary<string, JsonElement> _extensionData =
+        new Dictionary<string, JsonElement>();
+
+    /// <summary>
+    /// The reason code for why the transaction did not result in a reward
+    /// </summary>
+    [JsonPropertyName("reason")]
+    public required string Reason { get; set; }
+
+    /// <summary>
+    /// The display message associated to the notification
+    /// </summary>
+    [JsonPropertyName("message")]
+    public required string Message { get; set; }
+
+    /// <summary>
+    /// The transaction ID
+    /// </summary>
+    [JsonPropertyName("transactionId")]
+    public required string TransactionId { get; set; }
+
+    /// <summary>
+    /// The amount of the originating transaction in cents
+    /// </summary>
+    [JsonPropertyName("transactionAmountInCents")]
+    public required int TransactionAmountInCents { get; set; }
+
+    /// <summary>
+    /// The timestamp of the originating transaction in ISO format
+    /// </summary>
+    [JsonPropertyName("transactionTimestamp")]
+    public DateTime? TransactionTimestamp { get; set; }
+
+    [JsonIgnore]
+    public ReadOnlyAdditionalProperties AdditionalProperties { get; private set; } = new();
+
+    void IJsonOnDeserialized.OnDeserialized() =>
+        AdditionalProperties.CopyFromExtensionData(_extensionData);
+
+    /// <inheritdoc />
+    public override string ToString()
+    {
+        return JsonUtils.Serialize(this);
+    }
+}