+++ src/KardFinancial/Notifications/Types/RejectedTransactionRelationships.cs
@@ -1,0 +1,31 @@
+using global::System.Text.Json;
+using global::System.Text.Json.Serialization;
+using KardFinancial.Core;
+
+namespace KardFinancial;
+
+[Serializable]
+public record RejectedTransactionRelationships : IJsonOnDeserialized
+{
+    [JsonExtensionData]
+    private readonly IDictionary<string, JsonElement> _extensionData =
+        new Dictionary<string, JsonElement>();
+
+    [JsonPropertyName("user")]
+    public required RelationshipSingle User { get; set; }
+
+    [JsonPropertyName("transaction")]
+    public required RelationshipSingle Transaction { get; set; }
+
+    [JsonIgnore]
+    public ReadOnlyAdditionalProperties AdditionalProperties { get; private set; } = new();
+
+    void IJsonOnDeserialized.OnDeserialized() =>
+        AdditionalProperties.CopyFromExtensionData(_extensionData);
+
+    /// <inheritdoc />
+    public override string ToString()
+    {
+        return JsonUtils.Serialize(this);
+    }
+}